requirements.txt
selenium==3.141.0
requests==2.13.0
Markdown==3.1.1
PyAutoGUI==0.9.45
PyExecJS==1.5.1
pyperclip==1.7.0
Pillow==6.1.0
matplotlib==3.1.1
